requirements.txt
jupyterlab==4.1.6
pandas==2.2.2 
matplotlib==3.8.4
numpy==1.26.4
seaborn==0.13.2
scikit-learn==1.4.2 
scipy==1.13.0 
pillow==10.3.0
tensorflow==2.15.0
torch==1.13.1
imageio==2.34.0

